refactor(lint): widen token counts to u64; gate numeric casts at deny
Resolves every numeric-cast lint in the cleaned crates and locks the three
cast lints (cast_possible_truncation / sign_loss / possible_wrap) at deny so
they can never reappear silently — replacing the blanket workspace `allow`.

Token widening (the keystone — "fix the type, the cast vanishes"):
- Usage (llm), TokenCounts (protocol), ResponseLog (cache_forensics),
  ApiCallEntry (diagnostics), SessionTokens / DormantPingUsage (daemon),
  CallRow / pricing / query / cache_tracker (ledger) token fields: u32 → u64.
  Token sources are already u64 (API `as_u64`), so the narrowing casts in
  stream_helpers simply disappear.
- Fixes a latent truncation bug: the ledger read stored token columns as
  `i64 as u32`, silently truncating any cumulative/large value; now i64→u64
  via try_from.

Remaining casts, handled per the paragon playbook (no blanket allows):
- elapsed `Duration::as_millis()` (u128): saturating helpers in
  llm::convert (elapsed_ms_u32/u64) — u128 source means widening can't remove
  the cast, so an explicit saturating floor is the honest fix.
- breakpoint index arithmetic: checked usize/try_from, no `as`.
- FFI pid → pid_t: try_from with a Dead fallback (also fixes the old
  negative-pid wrap, where a bogus PID wrapped to a process-group id).
- f64→f32 embeddings / f64→u64 duration parse: scoped #[expect(reason=...)]
  (no try_from exists for float narrowing); #[expect] self-warns if it ever
  stops applying.

Cleaned crates now carry zero cast `as` and zero blanket allows for these
lints. Workspace compiles; ledger/llm/config/protocol/diagnostics/swp-* +
770 daemon tests pass.

+++ backend/daemon/src/autonomy/manager.rs
@@ -2271,8 +2271,8 @@ fn append_wrap_up_nudge(request: &mut LlmRequest) {
 // ---------------------------------------------------------------------------
 
 struct DormantPingUsage {
-    input_tokens: u32,
-    cache_read_tokens: u32,
+    input_tokens: u64,
+    cache_read_tokens: u64,
 }
 
 enum DormantPingOutcome {
